upgraded the DRS libraries to 5.0.1
Added support for board types 7 and 9  (only type 8 had been supported so far)

+++ daq_device_drs.h
@@ -51,7 +51,7 @@ private:
 
   int _trigger;
   int _trigger_handler;
-  float _tthreshold;
+  double _tthreshold;
   int _slope;
   int _delay;
   int _speed;
